createdAt 반영 (#6)
* 보안 및 코드 리펙터링 (#1)

* BE 프로젝트 초기 설정

* Rename Readme.md to README.md

* feat: 게시글 생성, 제거

* feat: 댓글 기능 추가

* feat: post

* Update PostController.java

* Update application.yml

* JWT 인증

* 시스템 로깅 제거

* 리펙터링

---------


* 중복 코드 수정

* createdAt 반영

---------

+++ be/src/main/java/org/example/be/post/PostService.java
@@ -7,6 +7,7 @@ import java.util.UUID;
 import org.example.be.post.dto.GetPostResponse;
 import org.example.be.post.dto.UpdatePostRequest;
 import org.example.be.post.dto.WritePostRequest;
+import org.example.be.user.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -20,7 +21,7 @@ public class PostService {
 
 	private final PostRepository postRepository;
 
-	public void writePost(WritePostRequest request, MultipartFile imageFile) throws IOException {
+	public void writePost(WritePostRequest request, MultipartFile imageFile, Member member) throws IOException {
 
 		String imagePath = System.getProperty("user.dir") + "\\src\\main\\resources\\static";
 		String fileName = imageFile != null ? imageFile.getOriginalFilename() : null;
@@ -39,14 +40,15 @@ public class PostService {
 			.content(request.content())
 			.imageName(newFileName)
 			.imagePath(newFileName != null ? imagePath + File.separator + newFileName : null)
+			.owner(member)
 			.build();
 
 		postRepository.save(post);
 	}
 
+	@Transactional(readOnly = true)
 	public GetPostResponse readPost(Long id) {
-		Post post = postRepository.findById(id)
-			.orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+		Post post = getPostById(id);
 
 		GetPostResponse.PostDetail postDetail = new GetPostResponse.PostDetail(post.getId(), post.getTitle(),
 			post.getContent(), post.getImagePath());
@@ -54,9 +56,10 @@ public class PostService {
 		return new GetPostResponse(postDetail);
 	}
 
-	public void updatePost(Long id, UpdatePostRequest request) {
-		Post post = postRepository.findById(id)
-			.orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+	public void updatePost(Long id, UpdatePostRequest request, Member member) {
+		Post post = getPostById(id);
+
+		validateMemberOwnership(member, post);
 
 		post.setTitle(request.title());
 		post.setContent(request.content());
@@ -64,10 +67,22 @@ public class PostService {
 		postRepository.save(post);
 	}
 
-	public void removePost(Long id) {
-		Post post = postRepository.findById(id)
-			.orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+	public void removePost(Long id, Member member) {
+		Post post = getPostById(id);
+
+		validateMemberOwnership(member, post);
 
 		postRepository.delete(post);
+	}
+
+	private Post getPostById(Long postId) {
+		return postRepository.findById(postId)
+			.orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다. ID: " + postId));
+	}
+
+	private void validateMemberOwnership(Member member, Post post) {
+		if (!post.getOwner().equals(member)) {
+			throw new IllegalArgumentException("게시글에 대한 권한이 없습니다.");
+		}
 	}
 }
